Consolidate pagination state and logic to hook
